backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv